+++ CHANGELOG.md
@@ -23,6 +23,7 @@
   - `setStockStrings()`
   - `exportSelfKeys()`
   - `importSelfKeys()`
+- breaking: jsonrpc: remove function `messageListGetMessageIds()`, it is replaced by `getMessageIds()` and `getMessageListEntries()` the latter returns a new `MessageListItem` type, which is the now prefered way of using the message list.
 - jsonrpc: add type: #3641, #3645
   - `MessageSearchResult`
   - `Location`